Move DataStore implementation to persistence-api
The existing `DataStore` implementation was implemented in each
persistence extension and was directly accessing methods of the
underlying store.

This commit instead create a single implemention of `DataStore` within
the `persistence-api` module which rely only on a `Persistence`
interface. This reduces the surface area of what a persistence extension
needs to implement, and also saves a lot of existing duplication
between extensions.

To simplify the changes, the patch also simplifies the `Persistence`
interface a bit.

Additionally, the patch simplifies the persistence extensions
implementations: we now directly generate native transport requests (so
Message.Request object) and call their execute() method directly. The
prior implementation was basically inlining the content of those
messages, but 1) this was more code and 2) this was a lot more fragile,
as any change in the logic of the messages in C*/DSE needed to be
reflected in Stargate extensions.