Remove initialize/destroy from Persistence interface
The `initialize` method was somewhat breaking the abstraction as it was
taking a config parameters, but there was no way in the `Persistence`
interface to generate such a config. So it was impossible to call that
initialize method without knowing the concrete implementation class.

Additionally, `initialize` should really be called at construction and
only once, so exposing it in the API in general did not add much but
could lead to error.

Overall, removing `initialize` (and `destroy`, which goes with it though
isn't even used anywhere) reduces the API in general, and remove one
of the type parameter.