Add Connection to the Persistence API
The ClientState wrapping done by the Persistence API was a smell of an
incomplete abstraction. Effectively, this was a cope-out for the fact
that queries are made against a "connection" which can have some state,
mainly a logged user and possibly some information on the connected
client.

As an aside, the fact `Persistence` had, as type argument, the
underlying persistence `ClientState` class, felt like another sign of
a leaky abstraction.

This commit makes this more explicit, cleaning (imo) the persistence API
and avoiding the previous class wrapping (which removes a little bit
of (boring) code in persistence extensions in the process).